requirements.txt
scapy
psycopg2-binary
requests
python-dotenv
